@@ -174,16 +179,16 @@
     "babel-plugin-macros": "^3.0.1",
     "babel-plugin-module-resolver": "^4.1.0",
     "camelcase": "6.3.0",
-    "eslint": "^8.21.0",
+    "eslint": "^8.23.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-custom-alias": "^1.3.0",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-ft-flow": "^2.0.1",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.8.2",
+    "eslint-plugin-jest": "^26.9.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
-    "firebase": "^9.9.2",
+    "firebase": "^9.9.4",
     "firebase-admin": "^11.0.1",
     "flat": "^5.0.2",
     "husky": "^8.0.1",
@@ -207,7 +212,7 @@
     "remote-redux-devtools": "^0.5.16",
     "rimraf": "^3.0.2",
     "ts-node": "^10.9.1",
-    "typescript": "^4.7.4",
+    "typescript": "^4.8.2",
     "world-countries": "^4.0.0"
   }
 }
