@@ -6,11 +6,11 @@ import User from '@model/user'
 import { littens, messages, users } from '@fixtures/tests'
 import { cleanFirestore } from '@utils/tests'
 
-afterAll(async () => {
-  await cleanFirestore()
-})
-
 describe('Performs an end-to-end user and litten journey', () => {
+  beforeAll(async () => {
+    await cleanFirestore()
+  })
+
   it(`creates ${users.length} new users`, async () => {
     const usersCollection = await User.collection.get()
     expect(usersCollection.size).toBe(0)
@@ -31,11 +31,15 @@ describe('Performs an end-to-end user and litten journey', () => {
     const littensCollection = await Litten.collection.get()
     expect(littensCollection.size).toBe(0)
 
-    for await (const litten of littens) {
-      litten.userUid = userUid
-      const newLitten = new Litten(litten)
-      await newLitten.create()
-    }
+    const littensToCreate = littens.map((litten) => {
+      const newLitten = new Litten({
+        ...litten,
+        userUid,
+      })
+      return newLitten.create()
+    })
+
+    await Promise.all(littensToCreate)
 
     const newLittensCollection = await Litten.collection.get()
     expect(newLittensCollection.size).toBe(littens.length)
@@ -135,7 +139,16 @@ describe('Performs an end-to-end user and litten journey', () => {
     const chatMessages = await newMessage.subscribeToChat().get()
     expect(chatMessages.size).toBe(1)
     const [newChatMessage] = chatMessages.docs
-    expect(new Message(newChatMessage)).toEqual(newMessage)
+    const newChatMessageData = {
+      ...newChatMessage.data(),
+      id: newChatMessage.id,
+    }
+    expect(new Message(newChatMessageData).toJSON()).toEqual(
+      expect.objectContaining({
+        ...newMessage.toJSON(),
+        metadata: expect.any(Object),
+      }),
+    )
     await newChat.get()
     expect(newChat.read).toHaveLength(1)
     expect(newChat.read).toContain(userUid)
