@@ -1,11 +1,8 @@
 export interface AuthSettings {
-  id?: string
-  photoURL?: string
-  callingCode?: string
-  country?: string
-  displayName?: string
-  email?: string
-  password?: string
-  phoneNumber?: string
-  [key: string]: unknown
+  id: string
+  photoURL: string
+  country: string
+  displayName: string
+  email: string
+  password: string
 }
